Optimized Sounds
ChatGPT Response to framedrop problem:
One way to potentially remedy the framerate drop when audio is playing is to optimize the audio files themselves. If the audio files are too large or are not compressed efficiently, it could put a strain on the system and cause performance issues. You can also try reducing the number of simultaneous sounds that are playing at any given time, or adjusting the volume levels of individual sounds to create a better balance. By preloading audio data, you can reduce the delay between the time an audio file is requested to play and the time it actually starts playing.