Implement "Extrovert" Achievement (Part 2)
Opting not to use uniqueInteractions set for now, until I find out what is going on with it. Just used a simple interactions++ and set the #interactions id element equal to it. The achievement works, but it only increases the interactions value when talking to the alien northwest from elon. It doesn't work on interacting with other aliens.